// src/app/interfaces/paciente.ts
export interface Paciente {
    idPaciente: number;
    nombre:     String;
    apellidos:  String;
    rut:        number;
    digVer:     String;
    direccion:  String;
    username:   String;
    email:      String;
    password:   String;
    idRegion:   number;
    idComuna:   number;
    pregunta:   String;
    respuesta:  String;
}